moved handler to src folder, split into 2 files and included method comments
