+++ package.json
@@ -2,7 +2,7 @@
   "name": "vatsim-data-handler",
   "version": "3.3.0",
   "description": "A data handler for the VATSIM status system",
-  "main": "handler.js",
+  "main": "src/methods.js",
   "scripts": {
     "test": "mocha --reporter spec",
     "cover": "node node_modules/istanbul/lib/cli.js cover node_modules/mocha/bin/_mocha -- -R spec test/*"
@@ -24,6 +24,7 @@
   },
   "homepage": "https://github.com/hcphoon01/VATSIM-Data-Handler#readme",
   "devDependencies": {
+    "babel-eslint": "^10.1.0",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "chai-json": "^1.0.0",
@@ -34,6 +35,7 @@
     "nodemon": "^1.19.3"
   },
   "dependencies": {
+    "eslint": "^6.8.0",
     "request": "^2.88.0"
   }
 }
